I#563 - Fix argument types in ENABLE_BROKEN_SPOOL code
Apparently this code (which by default is only built on Solaris)
was missed when the types were changed from gint to gsize in
commit 3fd3f0a82a5c636d5c48baf2c2997542dd1569f8.

Closes https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/563

+++ src/camel/camel-movemail.c
@@ -367,7 +367,7 @@ camel_movemail_copy_filter (gint fromfd,
 	gchar buffer[4096 + PRE_SIZE];
 	gint written = 0;
 	gchar *filterbuffer;
-	gint filterlen, filterpre;
+	gsize filterlen, filterpre;
 
 	d (printf ("writing %d bytes ... ", bytes));
 
@@ -487,7 +487,7 @@ camel_movemail_solaris (gint oldsfd,
 {
 	CamelMimeParser *mp;
 	gchar *buffer;
-	gint len;
+	gsize len;
 	gint sfd;
 	CamelMimeFilter *ffrom;
 	gint ret = 1;
